+++ TrabajoWeb/src/main/java/com/gamepride/platform/repository/IPersonRepository.java
@@ -1,6 +1,7 @@
 package com.gamepride.platform.repository;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -11,6 +12,9 @@ import com.gamepride.platform.model.Person;
 @Repository
 public interface IPersonRepository extends JpaRepository<Person, Long> {
 
-	@Query("SELECT e FROM Person e WHERE e.name LIKE %?1%")
-	List<Person> findPersonByName(String name);
+	
+	List<Person> findByName(String name);
+	
+	@Query("SELECT e FROM Person e left join fetch e.lancenters WHERE e.name LIKE %?1%")
+	Optional<Person> fetchByPersonIdWithLanCenters(Long id);
 }
