+++ TrabajoWeb/src/main/java/com/gamepride/platform/repository/IRoleRepository.java
@@ -9,8 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.gamepride.platform.model.Role;
 
 @Repository
-public interface IRolRepository extends JpaRepository<Role, Long> {
+public interface IRoleRepository extends JpaRepository<Role, Long> {
 
-	@Query("SELECT e FROM Role e WHERE e.name LIKE %?1%")
-	List<Role> findRoleByName(String name);
+	
 }
